Move generic cdns bindings to /pkg/apis
Also added magic comments for Zone and Record and the code generated from them.
