+++ pkg/apis/clouddns/v1/main_test.go
@@ -1,4 +1,4 @@
-package zone
+package v1
 
 import (
 	"testing"
