--- pkg/genericclouddns/genericclouddns.go
@@ -1,1 +1,0 @@
-package genericclouddns
